now correctly updates the end of line
(found by enabling/disabling concealcursor), make it so that
`did_wcol = true;` is computed when folds are concealed (otherwise
a bad value of concealcursor prevents it).
Interestingly, when on a concealed character, we need to move the amount
of concealed character.

TODO:
- display background for folded elements ?
- allow for multiple inline folds (not nested)
- fix fold indicator ? how ?

with the removal of fold_line, it would be nice if we could check
https://stackoverflow.com/questions/18275527/combining-vims-fold-and-conceal-feature
is working