test stuff
wtf am I doing this

wip

ok

Idea is to pass fold as ptr

so that we can have search highlights in fold too.
We are looking for a way to toggle between foldtext and line display.

Ideally we could remove or adapt the use of n_extra to be able to
display arabic characters everywhere.